pyproject.toml
[project]
name = "educanopos"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "peewee>=3.18.1",
    "prettyconf>=2.3.0",
    "pyyaml>=6.0.2",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "ipython>=9.3.0",
    "pytest>=8.4.1",
]
